refactor: move sync reader/writer into read/write modules
This is first step in a larger refactor to separate the IO types (Read,
Write, BufRead) into their own modules and provide implementations of
both encryption and decryption for all types. This commit is only a move
- no code has been changed except for `use` statements to eliminate
warnings.
